requirements.txt
flask
openai-whisper
transformers
gunicorn
streamlit
plotly